Scripts/requirements.txt
Flask==0.12.2
Flask-RESTful==0.3.6
Flask-Cors==3.0.3
Flask-Login==0.4.0
itsdangerous==0.24
Werkzeug==0.12.2
passlib==1.7.1
Flask-HTTPAuth==3.2.3
Flask-Cache==0.13.1
Flask-Compress==1.4.0
requests==2.21.0
bs4==0.0.1
xlrd==1.1.0
pandas==0.23.4
boto3==1.9.165
neomodel==3.3.1
sklearn
pytz==2017.2
six==1.10.0
joblib==0.10.3
mysql-connector
numpy==1.13.0
python-dateutil==2.6.1
pytz==2017.2
scikit-learn==0.18.1
scipy==0.19.0
SQLAlchemy==1.0.13
mysqlclient==1.3.7      
networkx==2.2   
